Using MVVM architecture
