+++ lib/features/home/model/product.dart
