+++ lib/features/home/view/widgets/specifications.dart
