+++ lib/utils/bottom_nav.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:iconly/iconly.dart';
-import 'package:salonika/screens/cart_page.dart';
-import 'package:salonika/screens/home_page.dart';
-import 'package:salonika/screens/profile_page.dart';
-import 'package:salonika/screens/services_page.dart';
+import 'package:salonika/features/cart/view/cart_page.dart';
+import 'package:salonika/features/home/view/home_page.dart';
+import 'package:salonika/features/profile/view/profile_page.dart';
+import 'package:salonika/features/services/view/services_page.dart';
 
 class BottomNav extends StatefulWidget {
   const BottomNav({super.key});
